RAG_app/requirements.txt
streamlit==1.33.0
langchain==0.2.1
langchain-community==0.2.1
langchain-huggingface==0.0.3
langchain-chroma==0.1.2
chromadb==0.5.0
sentence-transformers==2.7.0
pypdf==4.2.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1